fixed more erros of type casting

+++ src/com/CS4398/spc51/gods/gods/Zeus.java
@@ -57,10 +57,10 @@ public class Zeus extends God{
 	}
 	
 
-	public void reward(int previousBP, int currentBP, Believer believer)
+	public void reward(float previousBP, float currentBP, Believer believer)
 
 	{
-		switch(currentBP)
+		switch((int)currentBP)
 		{
 		case 10:  	Reward give = new GiveItem(believer.getPlayer(), "You have been good to me and my purpose");
 					give.execute();
